+++ src/frontend/hooks/index.ts
@@ -1,5 +1,6 @@
 export { default as useResourceEdit } from './use-resource-edit'
 export { default as useResourceNew } from './use-resource-new'
+export * from './use-selected-records'
 export * from './use-notice'
 export * from './use-translation'
 export * from './use-record'